requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pyairtable>=3.0.0
